Empty lines + attribute fixes
Empty line should be parsed but hidden in AST, or tests will break
(because there's an empty line after header).

Attributes are revamped: they should not be an essential part of title0,
and they can appear everywhere in the document. Moreover, attributes can
be unset with `!` sign before or after. So, attribute is now an extra
block type. Remove predefined attributes from grammar as nonsense.